Extract `AddFilterParameterMenu` component
