+++ frontend/src/metabase/dashboard/components/AddFilterParameterMenu/index.ts
@@ -1,0 +1,1 @@
+export * from "./AddFilterParameterMenu";
