+++ frontend/src/metabase/dashboard/components/DashboardHeader/buttons/AddFilterParameterButton/AddFilterParameterButton.tsx
@@ -1,4 +1,3 @@
-import { useLayoutEffect, useRef, useState } from "react";
 import { t } from "ttag";
 
 import { ToolbarButton } from "metabase/components/ToolbarButton";
@@ -7,38 +6,18 @@ import {
   hideAddParameterPopover,
   showAddParameterPopover,
 } from "metabase/dashboard/actions";
+import { AddFilterParameterMenu } from "metabase/dashboard/components/AddFilterParameterMenu";
 import { getIsAddParameterPopoverOpen } from "metabase/dashboard/selectors";
 import { useDispatch, useSelector } from "metabase/lib/redux";
 import { useRegisterShortcut } from "metabase/palette/hooks/useRegisterShortcut";
-import {
-  type ParameterSection,
-  getDashboardParameterSections,
-  getDefaultOptionForParameterSectionMap,
-} from "metabase/parameters/utils/dashboard-options";
-import { getParameterIconName } from "metabase/parameters/utils/ui";
-import { Icon, Menu, Text } from "metabase/ui";
+import type { ParameterMappingOptions } from "metabase-types/api";
 
 export const AddFilterParameterButton = () => {
-  const sections = getDashboardParameterSections();
-  const dispatch = useDispatch();
   const isOpened = useSelector(getIsAddParameterPopoverOpen);
-  const [rightSectionWidth, setRightSectionWidth] = useState(0);
-  const rightSectionWidthRef = useRef(0);
+  const dispatch = useDispatch();
 
-  const handleItemClick = (section: ParameterSection) => {
-    const defaultOption = getDefaultOptionForParameterSectionMap()[section.id];
-    if (defaultOption) {
-      dispatch(addParameter(defaultOption));
-    }
-  };
-
-  const handleRightSectionRef = (rightSection: HTMLDivElement | null) => {
-    if (rightSection) {
-      rightSectionWidthRef.current = Math.max(
-        rightSectionWidthRef.current,
-        rightSection.clientWidth,
-      );
-    }
+  const handleSelectOption = (option: ParameterMappingOptions) => {
+    dispatch(addParameter(option));
   };
 
   useRegisterShortcut(
@@ -54,55 +33,23 @@ export const AddFilterParameterButton = () => {
     [isOpened],
   );
 
-  useLayoutEffect(() => {
-    if (isOpened) {
-      setRightSectionWidth(rightSectionWidthRef.current);
-    }
-  }, [isOpened]);
-
   return (
-    <Menu
+    <AddFilterParameterMenu
       opened={isOpened}
-      onClose={() => dispatch(hideAddParameterPopover())}
       position="bottom-end"
-      trapFocus
+      onSelectOption={handleSelectOption}
+      onClose={() => dispatch(hideAddParameterPopover())}
     >
-      <Menu.Target>
-        <ToolbarButton
-          icon="filter"
-          onClick={() =>
-            isOpened
-              ? dispatch(hideAddParameterPopover())
-              : dispatch(showAddParameterPopover())
-          }
-          aria-label={t`Add a filter or parameter`}
-          tooltipLabel={t`Add a filter or parameter`}
-        />
-      </Menu.Target>
-      <Menu.Dropdown data-testid="add-filter-parameter-dropdown">
-        <Menu.Label>{t`Add a filter or parameter`}</Menu.Label>
-        {sections.map((section) => (
-          <Menu.Item
-            key={section.id}
-            leftSection={<Icon name={getParameterIconName(section.id)} />}
-            rightSection={
-              <Text
-                ref={handleRightSectionRef}
-                c="inherit"
-                miw={rightSectionWidth}
-              >
-                {section.description}
-              </Text>
-            }
-            aria-label={section.name}
-            onClick={() => handleItemClick(section)}
-          >
-            <Text c="inherit" fw="bold">
-              {section.name}
-            </Text>
-          </Menu.Item>
-        ))}
-      </Menu.Dropdown>
-    </Menu>
+      <ToolbarButton
+        icon="filter"
+        onClick={() =>
+          isOpened
+            ? dispatch(hideAddParameterPopover())
+            : dispatch(showAddParameterPopover())
+        }
+        aria-label={t`Add a filter or parameter`}
+        tooltipLabel={t`Add a filter or parameter`}
+      />
+    </AddFilterParameterMenu>
   );
 };
